# core/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: NeuroGate API
  description: |
    Enterprise AI Gateway API providing intelligent LLM routing, PII protection,
    semantic caching, and multi-provider orchestration.

    ## Features
    - **Multi-Provider Routing**: Intelligent routing across OpenAI, Anthropic, Google, AWS Bedrock, and Azure
    - **PII Protection**: Automatic detection and redaction of sensitive data (NeuroGuard)
    - **Semantic Caching**: 4-tier caching with semantic similarity matching
    - **Cost Tracking**: Real-time usage and cost analytics
    - **Agent Operations**: Trace and monitor AI agent workflows

    ## Authentication
    Use API key authentication via the `Authorization` header:
    ```
    Authorization: Bearer your-api-key
    ```

  version: 1.0.0
  contact:
    name: NeuroGate Team
    url: https://neurogate.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api.neurogate.dev
    description: Production
...
paths:
...
  /v1/chat/completions:
    post:
      tags: [Chat]
      summary: Create chat completion
      description: Send a chat completion request. Supports streaming via SSE when stream=true.
      operationId: createChatCompletion
      parameters:
        - name: X-Canary-Weight
          in: header
          description: Canary weight for A/B testing (0-100)
          schema:
            type: integer
        - name: X-Trace-Id
          in: header
          description: Trace ID for distributed tracing
          schema:
            type: string
        - name: X-Session-Id
          in: header
          description: Session ID for conversation tracking
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful completion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
            text/event-stream:
              schema:
                type: string
        '400':
          description: Invalid request
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error

  # ==================== Models API ====================
...
  /api/rag/strategy:
    post:
      tags: [RAG]
      summary: Determine RAG strategy
      operationId: determineStrategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Strategy determined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGStrategy'
  /api/rag/search:
    post:
      tags: [RAG]
      summary: Search documents
      operationId: searchDocuments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoredPoint'
  /api/rag/documents:
    post:
      tags: [RAG]
      summary: Add document
      operationId: addDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDocumentRequest'
      responses:
        '200':
          description: Document added
          content:
            application/json:
              schema:
                type: string

  # ==================== Cortex API ====================
...
  /api/prompts/commit:
    post:
      tags: [Prompts]
      summary: Commit prompt
      operationId: commitPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitRequest'
      responses:
        '200':
          description: Prompt committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptVersion'
  /api/prompts/versions:
    get:
      tags: [Prompts]
      summary: Get version history
      operationId: getVersionHistory
      parameters:
        - name: branchName
          in: query
          schema:
            type: string
            default: main
      responses:
        '200':
          description: History retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PromptVersion'
  /api/prompts/templates:
    post:
      tags: [Prompts]
      summary: Create template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateRequest'
      responses:
        '200':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptTemplate'
    get:
      tags: [Prompts]
      summary: Search templates
      operationId: searchTemplates
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Templates found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PromptTemplate'
components:
  schemas:
...
    ChatRequest:
      type: object
      required:
        - messages
      properties:
        model:
          type: string
          default: gpt-4o
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        temperature:
          type: number
          default: 0.7
        max_tokens:
          type: integer
        stream:
          type: boolean
          default: false
        tools:
          type: array
          items:
            type: object
        tool_choice:
          type: string
    ChatResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          default: chat.completion
        model:
          type: string
        choices:
          type: array
          items:
            $ref: '#/components/schemas/Choice'
        usage:
          $ref: '#/components/schemas/Usage'
        cache_hit:
          type: boolean
        route:
          type: string
    Message:
      type: object
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: string
        name:
          type: string
        tool_calls:
          type: array
          items:
            type: object
    Choice:
      type: object
      properties:
        index:
          type: integer
        message:
          $ref: '#/components/schemas/Message'
        finish_reason:
          type: string
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer

    # ==================== Models Schemas ====================
...
    RAGStrategy:
      type: object
      properties:
        strategy:
          type: string
        reason:
          type: string
        collections:
          type: array
          items:
            type: string
    SearchRequest:
      type: object
      properties:
        query:
          type: string
        limit:
          type: integer
        collection:
          type: string
    ScoredPoint:
      type: object
      properties:
        id:
          type: string
        score:
          type: number
        payload:
          type: object
    AddDocumentRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        source:
          type: string

    # ==================== Evaluation Schemas ====================
...
    CommitRequest:
      type: object
      properties:
        prompt_text:
          type: string
        commit_message:
          type: string
        author:
          type: string
        branch_name:
          type: string
    PromptVersion:
      type: object
      properties:
        version_id:
          type: string
        prompt_text:
          type: string
        commit_message:
          type: string
        author:
          type: string
        created_at:
          type: string
          format: date-time
    TemplateRequest:
      type: object
      properties:
        template_name:
          type: string
        description:
          type: string
        template_text:
          type: string
        variables:
          type: object
        tags:
          type: array
          items:
            type: string
    PromptTemplate:
      type: object
      properties:
        id:
          type: string
        template_name:
          type: string
        description:
          type: string
        template_text:
          type: string
        variables:
          type: object
        tags:
          type: array
          items:
            type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []